+++ pypeit/spectrographs/keck_hires.py
@@ -211,7 +211,6 @@ class KECKHIRESSpectrograph(spectrograph.Spectrograph):
         else:
             msgs.error("Not ready for this compound meta")
 
-
     def configuration_keys(self):
         """
         Return the metadata keys that define a unique instrument
@@ -226,7 +225,7 @@ class KECKHIRESSpectrograph(spectrograph.Spectrograph):
             and used to constuct the :class:`~pypeit.metadata.PypeItMetaData`
             object.
         """
-        return ['filter1', 'echangle', 'xdangle']
+        return ['filter1', 'echangle', 'xdangle', 'binning']
 
     def raw_header_cards(self):
         """