+++ pypeit/spectrographs/lbt_luci.py
@@ -43,7 +43,7 @@ class LBTLUCISpectrograph(spectrograph.Spectrograph):
 #                        use_darkimage=False)
 #        par.reset_all_processimages_par(**turn_off)
 #
-#        par['calibrations']['biasframe']['exprng'] = [None, 1]
+#        par['calibrations']['biasframe']['exprng'] = [None, 0.001]
 #        par['calibrations']['darkframe']['exprng'] = [999999, None]     # No dark frames
 #        par['calibrations']['pinholeframe']['exprng'] = [999999, None]  # No pinhole frames
 #        par['calibrations']['pixelflatframe']['exprng'] = [0, None]